+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/EntityPanel.java
@@ -44,7 +44,6 @@ import is.codion.swing.framework.ui.icon.FrameworkIcons;
 import javax.swing.AbstractAction;
 import javax.swing.JComponent;
 import javax.swing.JPanel;
-import javax.swing.JTabbedPane;
 import javax.swing.JToolBar;
 import javax.swing.KeyStroke;
 import javax.swing.SwingConstants;
@@ -58,8 +57,6 @@ import java.awt.Window;
 import java.awt.event.ActionEvent;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
-import java.beans.PropertyChangeEvent;
-import java.beans.PropertyChangeListener;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
@@ -68,7 +65,6 @@ import java.util.ResourceBundle;
 import java.util.function.Consumer;
 import java.util.function.Function;
 
-import static is.codion.swing.common.ui.Utilities.parentOfType;
 import static is.codion.swing.common.ui.Utilities.parentWindow;
 import static is.codion.swing.common.ui.component.Components.*;
 import static is.codion.swing.common.ui.component.button.ToggleButtonType.CHECKBOX;
@@ -266,23 +262,17 @@ public class EntityPanel extends JPanel {
   private final JPanel editControlTablePanel;
   private final Event<EntityPanel> activateEvent = Event.event();
   private final Value<String> caption;
+  private final Value<String> description;
   private final Value<PanelState> editPanelState = Value.value(EMBEDDED, EMBEDDED);
   private final State disposeEditDialogOnEscape = State.state(DISPOSE_EDIT_DIALOG_ON_ESCAPE.get());
 
   private final Config configuration;
 
-  private String description;
   private EntityPanel parentPanel;
   private EntityPanel previousSiblingPanel;
   private EntityPanel nextSiblingPanel;
 
   private boolean initialized = false;
-
-  static {
-    if (EntityEditPanel.USE_FOCUS_ACTIVATION.get()) {
-      KeyboardFocusManager.getCurrentKeyboardFocusManager().addPropertyChangeListener("focusOwner", new FocusActivationListener());
-    }
-  }
 
   /**
    * Instantiates a new EntityPanel instance. The panel is not laid out and initialized until {@link #initialize()} is called.
@@ -364,7 +354,7 @@ public class EntityPanel extends JPanel {
     this.entityModel = entityModel;
     String defaultCaption = entityModel.editModel().entityDefinition().caption();
     this.caption = Value.value(defaultCaption, defaultCaption);
-    this.description = entityModel.editModel().entityDefinition().description();
+    this.description = Value.value(entityModel.editModel().entityDefinition().description());
     this.editPanel = editPanel;
     this.tablePanel = tablePanel;
     this.editControlPanel = createEditControlPanel();
@@ -595,17 +585,10 @@ public class EntityPanel extends JPanel {
   }
 
   /**
-   * Sets the description text to use in f.ex. tool tips for tabbed panes
-   * @param description the description
+   * Setting this description Value to null reverts back to the default entity description.
+   * @return a Value for the description used when presenting this entity panel
    */
-  public final void setDescription(String description) {
-    this.description = description;
-  }
-
-  /**
-   * @return the description
-   */
-  public final String getDescription() {
+  public final Value<String> description() {
     return description;
   }
 
@@ -1202,42 +1185,7 @@ public class EntityPanel extends JPanel {
     }
 
     private Optional<EntityPanel> activeDetailPanel() {
-      Collection<EntityPanel> activeDetailPanels = activeDetailPanels();
-      if (!activeDetailPanels.isEmpty()) {
-        return Optional.of(activeDetailPanels.iterator().next());
-      }
-
-      return Optional.empty();
-    }
-  }
-
-  private static class FocusActivationListener implements PropertyChangeListener {
-    @Override
-    public void propertyChange(PropertyChangeEvent changeEvent) {
-      Component focusedComponent = (Component) changeEvent.getNewValue();
-      EntityPanel entityPanelParent = entityPanel(focusedComponent);
-      if (entityPanelParent != null) {
-        if (entityPanelParent.containsEditPanel()) {
-          entityPanelParent.editPanel().active().set(true);
-        }
-      }
-      else {
-        EntityEditPanel editPanelParent = parentOfType(EntityEditPanel.class, focusedComponent);
-        if (editPanelParent != null) {
-          editPanelParent.active().set(true);
-        }
-      }
-    }
-
-    private static EntityPanel entityPanel(Component focusedComponent) {
-      if (focusedComponent instanceof JTabbedPane) {
-        Component selectedComponent = ((JTabbedPane) focusedComponent).getSelectedComponent();
-        if (selectedComponent instanceof EntityPanel) {
-          return (EntityPanel) selectedComponent;
-        }
-      }
-
-      return parentOfType(EntityPanel.class, focusedComponent);
+      return activeDetailPanels().stream().findFirst();
     }
   }
 
@@ -1355,7 +1303,7 @@ public class EntityPanel extends JPanel {
     }
 
     /**
-     * @param includeControls true if the edit an table panel controls should be included
+     * @param includeControls true if the edit and table panel controls should be included
      * @return this Config instance
      */
     public Config includeControls(boolean includeControls) {