+++ changelog.md
@@ -8,6 +8,9 @@ Codion Change Log
 - TabbedDetailLayout, bunch of renaming due to recent changes.
 - DetailController merged with DetailLayout.
 - TabbedDetailLayout resizing bug fixed.
+- WindowDetailLayout added, used in SchemaBrowser demo.
+- EntityPanel.description() now Value based, for consistency.
+- EntityEditPanel.Config.focusActivation added, related refactoring.
 
 ## 0.17.32
 ### is.codion.common.core